// dashboard/src/client/models/ui-event-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const UiEventType = {
    RunStatus: 'run.status',
    RunSnapshot: 'run.snapshot',
    RunMetrics: 'run.metrics',
    NodeState: 'node.state',
    NodeStatus: 'node.status',
    NodeResultSnapshot: 'node.result.snapshot',
    NodeResultDelta: 'node.result.delta',
    NodeError: 'node.error',
    ArtifactReady: 'artifact.ready',
    ArtifactRemoved: 'artifact.removed',
    CommandAck: 'command.ack',
    CommandError: 'command.error',
    WorkerHeartbeat: 'worker.heartbeat',
    WorkerPackage: 'worker.package'
} as const;

export type UiEventType = typeof UiEventType[keyof typeof UiEventType];