// dashboard/src/client/models/run-status-event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { RunStatus } from './run-status';

export interface RunStatusEvent {
    'kind': RunStatusEventKindEnum;
    'runId': string;
    'status': RunStatus;
    'startedAt'?: string | null;
    'finishedAt'?: string | null;
    'reason'?: string | null;
}

export const RunStatusEventKindEnum = {
    RunStatus: 'run.status'
} as const;

export type RunStatusEventKindEnum = typeof RunStatusEventKindEnum[keyof typeof RunStatusEventKindEnum];