// dashboard/src/client/models/node-state-event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { WorkflowNodeState } from './workflow-node-state';

export interface NodeStateEvent {
    'kind': NodeStateEventKindEnum;
    'runId': string;
    'nodeId': string;
    'state': WorkflowNodeState;
}

export const NodeStateEventKindEnum = {
    NodeState: 'node.state'
} as const;

export type NodeStateEventKindEnum = typeof NodeStateEventKindEnum[keyof typeof NodeStateEventKindEnum];